+++ R/position-nudge-jitter.R
@@ -49,7 +49,7 @@
 #' @param x,y Amount of vertical and horizontal distance to move. A numeric
 #'   vector of length 1, or of the same length as rows there are in \code{data},
 #'   with nudge values in data rows order.
-#' @param direction One of \code{"as.is"}, code{"alternate"}, \code{"split"},
+#' @param direction One of \code{"as.is"}, \code{"alternate"}, \code{"split"},
 #'   \code{"split.x"} or \code{"split.y"}. A value of \code{"none"} replicates
 #'   the behavior of \code{\link[ggplot2]{position_nudge}}. \code{"split"}
 #'   changes the sign of the nudge depending on the direction of the random